change page structure
